requirements.txt
flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
supabase==1.0.4
pandas==2.0.3
uuid==1.30
python-dateutil==2.8.2
flask-cors==4.0.0 